web/requirements.txt
# Extra dependencies for the FastAPI + LangChain web server.
# Install with:  pip install -r web/requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0

# LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
